nlp-service/requirements.txt
# Framework web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# NLP e Machine Learning
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99

# Estrazione testo da documenti
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Utilità
requests==2.31.0
python-multipart==0.0.6
